Remove non handleExitError calls on BeforeFunc err
When the Before function of an app encountered an error, the
previous functionality was to write the error to the app's
default Writer without formatting and show the help text.
This is rigid and only precedes a call to the app's exit error
handler, whose behavior can be user defined.